Added sqlconsole()! This adds a sqlclient on the console (beware you need to type a single S before each sql line). Also you could type mil commands, ie need to be preceeded with a M.
Added a new table history. After a 'set explain='sql'; each statement
execution times, query text, user id are saved in this table.

Added version one of a logmanager (currently not activated). The logmanager
should take care of cleaning the bats in the central catalog and removing
log files which have been integrated into the current central bat set.

Added parsing of UNION/EXCEPT/INTERSECT.

Bug fixes for bugs reported by Martin and Fabian
crash on variants for cross and groupby (please add sql scripts to the
testing for your respective bugs).

Added check when ATOMfrom str didn't parse the string but returned a
correct value (nil) anyway. This is only usefull for MonetDB's external types,
such as blob, ie doesn't solve current bug report
(919531 Illegal casts not caught)

Compilation fixes for signed/unsiged, seems gcc on fedora is stricter as
older gcc's

+++ sql/src/odbc/samples/odbcsample1.c
@@ -17,7 +17,7 @@ prerr(SQLSMALLINT tpe, SQLHANDLE hnd, const char *func, const char *pref)
 
 	switch (SQLGetDiagRec(tpe, hnd, 1, state, &errnr, msg, sizeof(msg), &msglen)) {
 	case SQL_SUCCESS_WITH_INFO:
-		if (msglen >= sizeof(msg))
+		if (msglen >= (signed int)sizeof(msg))
 			fprintf(stderr, "(message truncated)\n");
 	case SQL_SUCCESS:
 		fprintf(stderr,
